# kvstore DB query use `BETWEEN` comparison, not `SUBSTR`, asymptotically better DB read perf