--- a/ledger/accountdb.go
+++ b/ledger/accountdb.go
@@ -4139,7 +4139,6 @@ func reencodeAccounts(ctx context.Context, tx *sql.Tx) (modifiedAccounts uint, e
 }
 
 // MerkleCommitter allows storing and loading merkletrie pages from a sqlite database.
-//
 //msgp:ignore MerkleCommitter
 type MerkleCommitter struct {
 	tx         *sql.Tx
@@ -4329,7 +4328,6 @@ func (iterator *encodedAccountsBatchIter) Close() {
 }
 
 // orderedAccountsIterStep is used by orderedAccountsIter to define the current step
-//
 //msgp:ignore orderedAccountsIterStep
 type orderedAccountsIterStep int