--- a/ledger/accountdb.go
+++ b/ledger/accountdb.go
@@ -2693,7 +2693,7 @@ func (qs *accountsDbQueries) lookupKeyValue(key string) (pv persistedKVData, err
 // The DB comparison for blob typed primary key is lexicographic, i.e., byte by byte.
 // In this way, we can introduce an interval that a primary key should be >= some prefix, < some prefix increment.
 // A corner case to consider is that, the prefix has last byte 0xFF, or the prefix is full of 0xFF.
-// The first case can be solved by carrying, e.g., prefix = 0x1EFF -> 0x1F00
+// The first case can be solved by carrying, e.g., prefix = 0x1EFF -> interval being >= 0x1EFF and < 0x1F
 // The second case can be solved by disregarding the upper limit, i.e., prefix = 0xFFFF -> interval being >= 0xFF
 // Another corner case to consider is empty byte, []byte{} or nil.
 // In both cases, the results are interval >= "", i.e., returns []byte{} for prefix, and nil for prefixIncr.
@@ -2706,7 +2706,7 @@ func keyPrefixIntervalPreprocessing(prefix []byte) ([]byte, []byte) {
 	for i := len(prefix) - 1; i >= 0; i-- {
 		currentByteIncr := int(prefix[i]) + 1
 		if currentByteIncr > 0xFF {
-			prefixIncr[i] = 0
+			prefixIncr = prefixIncr[:len(prefixIncr)-1]
 			continue
 		}
 		prefixIncr[i] = byte(currentByteIncr)