--- a/ledger/accountdb.go
+++ b/ledger/accountdb.go
@@ -2693,10 +2693,10 @@ func (qs *accountsDbQueries) lookupKeyValue(key string) (pv persistedKVData, err
 // The DB comparison for blob typed primary key is lexicographic, i.e., byte by byte.
 // In this way, we can introduce an interval that a primary key should be >= some prefix, < some prefix increment.
 // A corner case to consider is that, the prefix has last byte 0xFF, or the prefix is full of 0xFF.
-// The first case can be solved by carrying, e.g., prefix = 0x1EFF -> interval being >= 0x1EFF and < 0x1F
-// The second case can be solved by disregarding the upper limit, i.e., prefix = 0xFFFF -> interval being >= 0xFF
+// - The first case can be solved by carrying, e.g., prefix = 0x1EFF -> interval being >= 0x1EFF and < 0x1F
+// - The second case can be solved by disregarding the upper limit, i.e., prefix = 0xFFFF -> interval being >= 0xFFFF
 // Another corner case to consider is empty byte, []byte{} or nil.
-// In both cases, the results are interval >= "", i.e., returns []byte{} for prefix, and nil for prefixIncr.
+// - In both cases, the results are interval >= "", i.e., returns []byte{} for prefix, and nil for prefixIncr.
 func keyPrefixIntervalPreprocessing(prefix []byte) ([]byte, []byte) {
 	if prefix == nil {
 		prefix = []byte{}