--- a/ledger/accountdb_test.go
+++ b/ledger/accountdb_test.go
@@ -532,11 +532,12 @@ func checkCreatables(t *testing.T,
 // randomCreatableSampling sets elements to delete from previous iteration
 // It consideres 10 elements in an iteration.
 // loop 0: returns the first 10 elements
-// loop 1: returns: * the second 10 elements
-//   - random sample of elements from the first 10: created changed from true -> false
-//
-// loop 2: returns: * the elements 20->30
-//   - random sample of elements from 10->20: created changed from true -> false
+// loop 1: returns:
+// - the second 10 elements
+// - random sample of elements from the first 10: created changed from true -> false
+// loop 2: returns:
+// - the elements 20->30
+// - random sample of elements from 10->20: created changed from true -> false
 func randomCreatableSampling(iteration int, crtbsList []basics.CreatableIndex,
 	creatables map[basics.CreatableIndex]ledgercore.ModifiedCreatable,
 	expectedDbImage map[basics.CreatableIndex]ledgercore.ModifiedCreatable,
@@ -3239,12 +3240,11 @@ func BenchmarkBoxDatabaseRead(b *testing.B) {
 //
 // addr | rnd | status
 // -----|-----|--------
-//
-//	A |   1 |      1
-//	B |   1 |      1
-//	A |   2 |      0
-//	B |   3 |      0
-//	C |   3 |      1
+// A    |  1  |    1
+// B    |  1  |    1
+// A    |  2  |    0
+// B    |  3  |    0
+// C    |  3  |    1
 //
 // Ensure
 // - for round 1 A and B returned