<!--
Thanks for submitting a pull request! We appreciate the time and effort you spent to get this far.

If you haven't already, please make sure that you've reviewed the CONTRIBUTING guide:
https://github.com/algorand/go-algorand/blob/master/CONTRIBUTING.md#code-guidelines

In particular ensure that you've run the following:
* make generate
* make sanity (which runs make fmt, make lint, make fix and make vet)

It is also a good idea to run tests:
* make test
* make integration
-->

## Summary

<!-- Explain the goal of this change and what problem it is solving. Format this cleanly so that it may be used for a commit message, as your changes will be squash-merged. -->

Previous:

```
go test -benchmem -run=^$ -bench ^BenchmarkLookupKeyByPrefix$ github.com/algorand/go-algorand/ledger
goos: darwin
goarch: arm64
pkg: github.com/algorand/go-algorand/ledger
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize4-10         	  295464	      4067 ns/op	     992 B/op	      42 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize16-10        	  245655	      4754 ns/op	     992 B/op	      42 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize64-10        	  158940	      7450 ns/op	     992 B/op	      42 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize256-10       	   66130	     18040 ns/op	     992 B/op	      42 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize1024-10      	   19956	     60110 ns/op	     992 B/op	      42 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize4096-10      	    4958	    236905 ns/op	     992 B/op	      42 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize16384-10     	    1286	    921900 ns/op	     992 B/op	      42 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize65536-10     	     325	   3663356 ns/op	     992 B/op	      42 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize262144-10    	      76	  15481009 ns/op	     992 B/op	      42 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize1048576-10   	      16	  63826320 ns/op	     993 B/op	      42 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize4194304-10   	       4	 259755125 ns/op	     992 B/op	      42 allocs/op
PASS
ok  	github.com/algorand/go-algorand/ledger	44.107s
```

Now:

```
❯ go test -benchmem -run=^$ -bench ^BenchmarkLookupKeyByPrefix$
goos: darwin
goarch: arm64
pkg: github.com/algorand/go-algorand/ledger
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize4-10         	  291423	      4117 ns/op	    1176 B/op	      48 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize16-10        	  281562	      4139 ns/op	    1176 B/op	      48 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize64-10        	  279552	      4150 ns/op	    1176 B/op	      48 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize256-10       	  280502	      4181 ns/op	    1176 B/op	      48 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize1024-10      	  276249	      4200 ns/op	    1176 B/op	      48 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize4096-10      	  276704	      4216 ns/op	    1176 B/op	      48 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize16384-10     	  273567	      4241 ns/op	    1176 B/op	      48 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize65536-10     	  262652	      4267 ns/op	    1176 B/op	      48 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize262144-10    	  265056	      4276 ns/op	    1176 B/op	      48 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize1048576-10   	  267559	      4298 ns/op	    1176 B/op	      48 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize4194304-10   	  250052	      4332 ns/op	    1176 B/op	      48 allocs/op
PASS
ok  	github.com/algorand/go-algorand/ledger	43.417s
```

From $O(n)$ read to DB to $O(\log n)$ binary search.

The full benchmark takes about 3 hours to run, DB size grow from 2 to 2^27.

```
❯ go test -benchmem -run=^$ -bench ^BenchmarkLookupKeyByPrefix$ -timeout 99999s

goos: darwin
goarch: arm64
pkg: github.com/algorand/go-algorand/ledger
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize2-10         	  238437	      4200 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize4-10         	  277704	      4186 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize8-10         	  276018	      4203 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize16-10        	  274482	      4209 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize32-10        	  275900	      4219 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize64-10        	  275570	      4222 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize128-10       	  271407	      4246 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize256-10       	  270850	      4253 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize512-10       	  271416	      4267 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize1024-10      	  271671	      4262 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize2048-10      	  270013	      4284 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize4096-10      	  269492	      4285 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize8192-10      	  268980	      4302 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize16384-10     	  266671	      4317 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize32768-10     	  267967	      4343 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize65536-10     	  258992	      4357 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize131072-10    	  256395	      4362 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize262144-10    	  260775	      4368 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize524288-10    	  261010	      4386 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize1048576-10   	  258439	      4403 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize2097152-10   	  257515	      4396 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize4194304-10   	  245020	      4442 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize8388608-10   	  227670	      4435 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize16777216-10  	  271060	      4431 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize33554432-10  	  269516	      4445 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize67108864-10  	  256909	      4476 ns/op	    1056 B/op	      45 allocs/op
BenchmarkLookupKeyByPrefix/lookupKVByPrefix-DBsize134217728-10 	  247914	      4542 ns/op	    1056 B/op	      45 allocs/op
PASS
ok  	github.com/algorand/go-algorand/ledger	11922.805s
```

## Test Plan

<!-- How did you test these changes? Please provide the exact scenarios you tested in as much detail as possible including commands, output and rationale. -->

Some corner case discussed with @jannotti need to be implemented, when prefix is all `0xFF` bytes.